# Improve performance of placeholder and add ghs performance test