--- a/tests/_data/data-sets/index.js
+++ b/tests/_data/data-sets/index.js
@@ -3,14 +3,22 @@
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-licensing-options
  */
 
-import paragraphs from './paragraphs.js';
-import lists from './lists.js';
-import tableHuge from './table-huge.js';
 import formattingLongP from './formatting-long-paragraphs.js';
+import ghs from './ghs.js';
 import inlineStyles from './inline-styles.js';
+import lists from './lists.js';
 import mixed from './mixed.js';
+import paragraphs from './paragraphs.js';
+import tableHuge from './table-huge.js';
 import wiki from './wiki.js';
 
 export default {
-	paragraphs, lists, tableHuge, formattingLongP, inlineStyles, mixed, wiki
+	formattingLongP,
+	ghs,
+	inlineStyles,
+	lists,
+	mixed,
+	paragraphs,
+	tableHuge,
+	wiki
 };