### Suggested merge commit message ([convention](https://ckeditor.com/docs/ckeditor5/latest/framework/contributing/git-commit-message-convention.html))

Other (engine): Improve performance of the placeholders.

Other (image): Attribute `loading="lazy"` will be automatically added in editing view to images with `height` and `width` attributes set to improve loading performance.

Tests: Added new `ghs` manual performance test.

MINOR BREAKING CHANGE (image): Starting from this release, images that have `height` and `width` attributes set will automatically receive `loading="lazy"` attribute in the editing area. This happens only for the content loaded into the editor, the data output produced by the editor remains the same. The reason for this change is to improve user experience in documents that may contain hundreds of images.

---

This is a continuation of #17589.

When looking at the numbers reported by the performance tests, we can see that these changes have reduced the load time of the `ghs` test by around 23% (from 9400ms to 7300ms).

However, when images were present in the document, a lot of code was run after they were loaded, so the browser window was still frozen for a few seconds after the editor was rendered. When this is considered, the performance is now 2 times better (loading, scripting, rendering, and painting took 17800ms before, but only 8900 ms after, including devtools overhead).